fix(issue#4): use unicode strings and µ for micro
Prior to this commit, return types were `str` and the character `u` was
used to denote the SI unit prefix for *"micro"*.  However, as discussed in
issue #4, the Bureau International des Poids et Mesures specifies that the
character `Â` (i.e., `u'\N{MICRO SIGN}'`, `'\x0b5'`) **MUST** be used to
denote the SI unit prefix for *"micro"*.  See [this report][1] for
details.

In this commit:

 - Define SI prefix units as unicode strings.
 - Specify SI unit portion of regular expressions as unicode.
 - Change return types to unicode strings.

Fixes issue #4.

[1]: https://www.bipm.org/utils/common/pdf/si_brochure_8_en.pdf