ci: replace setuptools_scm write_to with version_file
Prior to this commit, the location of the version file generated by `setuptools_scm` was
specified using the `write_to` property. However, this seems to result in the version
resolving as "unknown" from `si_prefix.__version__`.

In this commit, as per the [setuptools_scm docs][1], set the version file location
using `version_file` instead.

[1]: https://github.com/pypa/setuptools_scm?tab=readme-ov-file#pyprojecttoml-usage